% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{apply_perspective}
\alias{apply_perspective}
\title{Apply perspective effect to 3D points}
\usage{
apply_perspective(rotated, dist)
}
\arguments{
\item{rotated}{Matrix of rotated x, y, z coordinates}

\item{dist}{Distance from viewer to center of scene}
}
\value{
Matrix with perspective effect applied
}
\description{
Apply perspective effect to 3D points
}
