% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting.R
\name{apply_surface_light}
\alias{apply_surface_light}
\title{Apply lighting models to surface normals}
\usage{
apply_surface_light(face_data, normals, face_centers, light)
}
\arguments{
\item{face_data}{Data frame with unique faces (group column)}

\item{normals}{Matrix of surface normals}

\item{face_centers}{Matrix of face centers}

\item{light}{Lighting specification object}
}
\value{
Data frame with lighting values and normal components
}
\description{
Apply lighting models to surface normals
}
\keyword{internal}
