% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{attach_light}
\alias{attach_light}
\title{Store lighting specification to data frame for downstream use}
\usage{
attach_light(data, light)
}
\arguments{
\item{data}{Data frame to attach lighting spec to}

\item{light}{Lighting specification object, "none", or NULL}
}
\value{
Data frame with lighting_spec column added (if light is not NULL)
}
\description{
Store lighting specification to data frame for downstream use
}
\keyword{internal}
