% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{blend_light_with_colors}
\alias{blend_light_with_colors}
\title{Apply shading to colors based on lighting values}
\usage{
blend_light_with_colors(colors, light_values, lighting)
}
\arguments{
\item{colors}{Vector of color values (hex or named colors)}

\item{light_values}{Numeric vector of lighting values}

\item{lighting}{Lighting specification with shade_mode and shade_strength}
}
\value{
Vector of modified colors
}
\description{
Apply shading to colors based on lighting values
}
\keyword{internal}
