% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-polygon-3d.R
\name{blend_lighting_with_colors}
\alias{blend_lighting_with_colors}
\title{Blend lighting values with base colors using HSV or HSL color spaces}
\usage{
blend_lighting_with_colors(base_colors, light_values, lighting)
}
\description{
Blend lighting values with base colors using HSV or HSL color spaces
}
