% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{calculate_face_centers}
\alias{calculate_face_centers}
\title{Calculate face centers from vertex data}
\usage{
calculate_face_centers(data)
}
\arguments{
\item{data}{Data frame with x, y, z coordinates and group column}
}
\value{
Matrix with 3 columns (x, y, z face center coordinates)
}
\description{
Calculate face centers from vertex data
}
\keyword{internal}
