% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-pillar-3d.R
\name{calculate_pillar_face_normals}
\alias{calculate_pillar_face_normals}
\title{Calculate normals for pillar faces}
\usage{
calculate_pillar_face_normals(pillar_faces)
}
\arguments{
\item{pillar_faces}{Data frame with pillar face vertices}
}
\value{
Matrix of face normals (one row per face, 3 columns for x,y,z)
}
\description{
Calculate normals for pillar faces
}
