% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-surface-3d.R
\name{calculate_surface_face_centers}
\alias{calculate_surface_face_centers}
\title{Calculate face centers for positional lighting}
\usage{
calculate_surface_face_centers(faces, face_data)
}
\arguments{
\item{faces}{Data frame with all face vertices}

\item{face_data}{Data frame with unique faces}
}
\value{
Matrix with face centers (one row per face, 3 columns)
}
\description{
Calculate face centers for positional lighting
}
\keyword{internal}
