% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-voxel.R
\name{calculate_voxel_face_centers}
\alias{calculate_voxel_face_centers}
\title{Calculate face centers for voxel faces}
\usage{
calculate_voxel_face_centers(voxel_faces)
}
\arguments{
\item{voxel_faces}{Data frame with voxel face vertices}
}
\value{
Matrix of face centers (one row per face, 3 columns for x,y,z)
}
\description{
Calculate face centers for voxel faces
}
