% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-voxel-3d.R
\name{calculate_voxel_face_normals}
\alias{calculate_voxel_face_normals}
\title{Calculate normals for voxel faces}
\usage{
calculate_voxel_face_normals(voxel_faces)
}
\arguments{
\item{voxel_faces}{Data frame with voxel face vertices}
}
\value{
Matrix of face normals (one row per face, 3 columns for x,y,z)
}
\description{
Calculate normals for voxel faces
}
