% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bar-3d.R
\name{compute_bar_stats}
\alias{compute_bar_stats}
\title{Compute normalized bar statistics}
\usage{
compute_bar_stats(agg_data)
}
\arguments{
\item{agg_data}{Data frame with count and bin_area columns}
}
\value{
Data frame with added proportion, ncount, density, ndensity columns
}
\description{
Compute normalized bar statistics
}
\keyword{internal}
