% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bar-3d.R
\name{compute_bin_2d}
\alias{compute_bin_2d}
\title{Bin continuous x and y data into 2D grid}
\usage{
compute_bin_2d(data, bins, binwidth, drop = TRUE)
}
\arguments{
\item{data}{Data frame with x, y, weight columns}

\item{bins}{Vector of length 2 giving number of bins in x and y}

\item{binwidth}{Vector of length 2 giving bin widths (overrides bins if provided)}

\item{drop}{Whether to drop empty bins}
}
\value{
Data frame with x, y, count, bin_area columns
}
\description{
Bin continuous x and y data into 2D grid
}
\keyword{internal}
