% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bar-3d.R
\name{compute_bin_mixed}
\alias{compute_bin_mixed}
\title{Bin mixed discrete/continuous data}
\usage{
compute_bin_mixed(data, bins, binwidth, drop = TRUE, x_discrete)
}
\arguments{
\item{data}{Data frame with x, y, weight columns}

\item{bins}{Vector of length 2 giving number of bins}

\item{binwidth}{Vector of length 2 giving bin widths}

\item{drop}{Whether to drop empty bins}

\item{x_discrete}{Logical indicating if x is discrete (y is continuous) or vice versa}
}
\value{
Data frame with x, y, count, bin_area columns
}
\description{
Bin mixed discrete/continuous data
}
\keyword{internal}
