% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bar-3d.R
\name{compute_count_2d}
\alias{compute_count_2d}
\title{Count discrete x-y combinations}
\usage{
compute_count_2d(data, drop = TRUE)
}
\arguments{
\item{data}{Data frame with x, y, weight columns}

\item{drop}{Whether to drop empty combinations}
}
\value{
Data frame with x, y, count, bin_area columns
}
\description{
Count discrete x-y combinations
}
\keyword{internal}
