% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting.R
\name{compute_light_dot_products}
\alias{compute_light_dot_products}
\title{Compute dot products between normals and light direction}
\usage{
compute_light_dot_products(normals, light_dir_norm)
}
\arguments{
\item{normals}{Matrix with 3 columns (x, y, z normal components)}

\item{light_dir_norm}{Normalized light direction vector}
}
\value{
Vector of dot products
}
\description{
Compute dot products between normals and light direction
}
