% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{compute_surface_normals}
\alias{compute_surface_normals}
\title{Compute surface normals from gradients}
\usage{
compute_surface_normals(gradients)
}
\arguments{
\item{gradients}{Data frame with group, dzdx, dzdy columns}
}
\value{
Matrix with 3 columns (x, y, z normal components)
}
\description{
Compute surface normals from gradients
}
\keyword{internal}
