% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-surface-3d.R
\name{convert_to_quads}
\alias{convert_to_quads}
\title{Create quadrilateral faces from grid data}
\usage{
convert_to_quads(data)
}
\arguments{
\item{data}{Regular grid data frame}
}
\value{
Data frame with quad faces and computed gradients
}
\description{
Create quadrilateral faces from grid data
}
\keyword{internal}
