% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-surface-3d.R
\name{create_grid_quads}
\alias{create_grid_quads}
\title{Create quadrilateral faces from grid data}
\usage{
create_grid_quads(data, light)
}
\arguments{
\item{data}{Regular grid data frame}
}
\value{
Data frame with quad faces and computed gradients
}
\description{
Create quadrilateral faces from grid data
}
\keyword{internal}
