% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-rendering.R
\name{create_grid_segments}
\alias{create_grid_segments}
\title{Create grid segments for rendering}
\usage{
create_grid_segments(grid_data, plot_bounds)
}
\arguments{
\item{grid_data}{Transformed grid data}

\item{plot_bounds}{Bounds for the final plot \link{xmin, xmax, ymin, ymax}}
}
\value{
Data frame with segment coordinates for grid grob
}
\description{
Create grid segments for rendering
}
\keyword{internal}
