% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colorbar_shaded.R
\name{create_lighting_gradients}
\alias{create_lighting_gradients}
\title{Create shading gradients for colorbar}
\usage{
create_lighting_gradients(
  base_colors,
  lighting_info,
  shade_range = c(-1, 1),
  n_steps = 20
)
}
\arguments{
\item{base_colors}{Character vector of base colors}

\item{lighting_info}{Lighting specification object}

\item{n_steps}{Number of lighting gradient steps (default 10)}
}
\value{
Matrix of colors with lighting applied
}
\description{
Create shading gradients for colorbar
}
\keyword{internal}
