% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{delaunay_points_to_tiles}
\alias{delaunay_points_to_tiles}
\title{Convert irregular points to triangular tiles via Delaunay triangulation}
\usage{
delaunay_points_to_tiles(data, group_prefix)
}
\description{
Convert irregular points to triangular tiles via Delaunay triangulation
}
\keyword{internal}
