% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{detect_scale_direction_flip}
\alias{detect_scale_direction_flip}
\title{Detect if a scale transformation flips direction}
\usage{
detect_scale_direction_flip(scale_obj)
}
\arguments{
\item{scale_obj}{A ggplot2 scale object}
}
\value{
Logical indicating if the scale transform flips direction
}
\description{
Detect if a scale transformation flips direction
}
\keyword{internal}
