% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-function-3d.R
\name{ensure_nonempty_data}
\alias{ensure_nonempty_data}
\title{Ensure non-empty data for stats that generate their own data}
\usage{
ensure_nonempty_data(data)
}
\description{
Ensure non-empty data for stats that generate their own data
}
\keyword{internal}
