% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{extract_aesthetic_vars}
\alias{extract_aesthetic_vars}
\title{Extract variable names from aesthetic mappings}
\usage{
extract_aesthetic_vars(plot_obj)
}
\arguments{
\item{plot_obj}{A ggplot object}
}
\value{
A list with x, y, z character vectors of variable names
}
\description{
Extract variable names from aesthetic mappings
}
