% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_3d.R
\name{extract_light_from_plot}
\alias{extract_light_from_plot}
\title{Extract lighting information from plot layers}
\usage{
extract_light_from_plot()
}
\value{
List with lighting specification or NULL if none found
}
\description{
Extract lighting information from plot layers
}
\keyword{internal}
