% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-3d.R
\name{facing_to_normal}
\alias{facing_to_normal}
\title{Convert facing specification to normal vector}
\usage{
facing_to_normal(facing)
}
\arguments{
\item{facing}{Character string ("xmin", "xmax", "ymin", "ymax", "zmin", "zmax")
or numeric vector of length 3}
}
\value{
Normalized vector of length 3
}
\description{
Convert facing specification to normal vector
}
\keyword{internal}
