% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{generate_circle_vertices}
\alias{generate_circle_vertices}
\title{Generate circle vertices in 3D space for a given face}
\usage{
generate_circle_vertices(x_std, y_std, z_std, face, radius, n_vertices)
}
\arguments{
\item{x_std, y_std, z_std}{Standardized point coordinates}

\item{face}{Face name (e.g., "zmin", "xmax")}

\item{radius}{Circle radius in standardized units}

\item{n_vertices}{Number of vertices for the circle}
}
\value{
Data frame with x, y, z coordinates for circle vertices
}
\description{
Generate circle vertices in 3D space for a given face
}
\keyword{internal}
