% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcube-package.R
\docType{package}
\name{ggcube-package}
\alias{ggcube}
\alias{ggcube-package}
\title{ggcube: 3D plotting with ggplot2}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewkling/ggcube}
  \item Report bugs at \url{https://github.com/matthewkling/ggcube/issues}
}

}
\keyword{internal}
