% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcube-package.R
\docType{package}
\name{ggcube}
\alias{ggcube-package}
\alias{ggcube}
\title{ggcube: Extend ggplot2 into 3D}
\description{
ggcube provides 3D geoms, stats, and coordinate systems for ggplot2.
Add \code{\link[=coord_3d]{coord_3d()}} to any plot with a z aesthetic to render it in 3D
with rotation, perspective, lighting, and depth sorting. Layer types
include surfaces, hulls, contours, ridgelines, paths, points, segments,
text, bars, and columns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewkling/ggcube}
  \item \url{https://matthewkling.github.io/ggcube/}
  \item Report bugs at \url{https://github.com/matthewkling/ggcube/issues}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mattkling@berkeley.edu} [copyright holder]

}
\keyword{internal}
