% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{grid_params}
\alias{grid_params}
\title{Grid generation parameters}
\arguments{
\item{grid, n, direction, trim}{Parameters determining the geometry, resolution, and
orientation of the surface grid. See \link{grid_generation} for details.}
}
\description{
Grid generation parameters
}
\keyword{internal}
