% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{grid_points_to_tiles}
\alias{grid_points_to_tiles}
\title{Convert regular grid points to tiles}
\usage{
grid_points_to_tiles(data, grid_type, group_prefix)
}
\description{
Convert regular grid points to tiles
}
\keyword{internal}
