% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-polygon-3d.R
\name{hsl2rgb}
\alias{hsl2rgb}
\title{Convert HSL to RGB color space}
\usage{
hsl2rgb(hsl_matrix)
}
\arguments{
\item{hsl_matrix}{3xN matrix with HSL values (H in \link{0,1}, S in \link{0,1}, L in \link{0,1})}
}
\value{
3xN matrix with RGB values in \link{0,1} range
}
\description{
Convert HSL to RGB color space
}
