% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{is_theme_void_like}
\alias{is_theme_void_like}
\title{Check if theme is void-like (has multiple key elements set to element_blank)}
\usage{
is_theme_void_like(theme_obj)
}
\arguments{
\item{theme_obj}{Theme object or theme list}
}
\value{
Logical indicating if theme appears to be void-like
}
\description{
Check if theme is void-like (has multiple key elements set to element_blank)
}
\keyword{internal}
