% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{light_param}
\alias{light_param}
\title{Light param}
\arguments{
\item{light}{A lighting specification object created by \code{\link[=light]{light()}}, \code{light("none")}
or the string \code{"none"} to disable lighting, or \code{NULL} to inherit plot-level
lighting. Set plot-level lighting by adding \code{\link[=light]{light()}} to the plot, and
layer-specific lighting via the \code{light} argument of \verb{geom_*_3d()} functions,
which takes precedence.}
}
\description{
Light param
}
\keyword{internal}
