% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-path-3d.R
\name{lorenz_attractor}
\alias{lorenz_attractor}
\title{Lorenz butterfly attractor generator}
\usage{
lorenz_attractor(n_points = 5000, dt = 0.01, sigma = 10, rho = 28, beta = 8/3)
}
\description{
Lorenz butterfly attractor generator
}
