% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-3d.R
\name{make_annotation_point_rows}
\alias{make_annotation_point_rows}
\title{Generate point annotation rows}
\usage{
make_annotation_point_rows(spec, group_offset)
}
\description{
Generate point annotation rows
}
\keyword{internal}
