% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-3d.R
\name{make_annotation_segment_rows}
\alias{make_annotation_segment_rows}
\title{Generate segment annotation rows}
\usage{
make_annotation_segment_rows(spec, group_offset)
}
\description{
Generate segment annotation rows
}
\keyword{internal}
