% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-3d.R
\name{make_annotation_text_rows}
\alias{make_annotation_text_rows}
\title{Generate text annotation rows}
\usage{
make_annotation_text_rows(spec, group_offset)
}
\description{
Generate text annotation rows
}
\keyword{internal}
