% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{make_point_grid}
\alias{make_point_grid}
\title{Generate a grid of unique vertex points}
\usage{
make_point_grid(
  grid = c("rectangle", "right1", "right2", "equilateral"),
  n = 40,
  direction = c("x", "y"),
  xlim,
  ylim,
  trim = TRUE
)
}
\arguments{
\item{grid, n, direction, trim}{Parameters determining the geometry, resolution, and
orientation of the surface grid. See \link{grid_generation} for details.}

\item{xlim, ylim}{Length-two numeric vectors defining bounding box.}
}
\value{
A data frame with columns \code{x}, \code{y}.
}
\description{
Creates a regular grid of unique vertex positions for use by surface-generating
stats. The grid is tessellated into polygon tiles by \code{\link[=points_to_tiles]{points_to_tiles()}} in
the geom layer (or in the stat for \code{\link[=stat_smooth_3d]{stat_smooth_3d()}}).
}
\keyword{internal}
