% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mountain}
\alias{mountain}
\title{Mountain}
\format{
\subsection{\code{mountain}}{

A data frame with 1290 rows and 3 columns (x, y, and z)
}
}
\source{
derived from the \code{datasets::volcano} data set
}
\usage{
mountain
}
\description{
A terrain data set representing elevation variation over
an x-y grid.
}
\keyword{datasets}
