% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting.R
\name{normalize_light_direction}
\alias{normalize_light_direction}
\title{Normalize light direction vector}
\usage{
normalize_light_direction(light_dir)
}
\arguments{
\item{light_dir}{Numeric vector of length 3}
}
\value{
Normalized light direction vector
}
\description{
Normalize light direction vector
}
\keyword{internal}
