% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{points_to_circles}
\alias{points_to_circles}
\title{Convert ref_circle points to circular polygons}
\usage{
points_to_circles(data, data_std)
}
\arguments{
\item{data}{Original data frame}

\item{data_std}{Standardized data frame}
}
\value{
Data frame with circular polygons replacing ref_circle points
}
\description{
Convert ref_circle points to circular polygons
}
\keyword{internal}
