% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{points_to_tiles}
\alias{points_to_tiles}
\title{Convert point grid to polygon tiles}
\usage{
points_to_tiles(
  data,
  method = "auto",
  grid_type = "rectangle",
  group_prefix = "surface__tile"
)
}
\arguments{
\item{data}{Data frame with x, y, z columns (and optionally row, column).}

\item{method}{Tessellation method: "grid" for regular grid, "delaunay" for
triangulation, "auto" to detect.}

\item{grid_type}{For method="grid", type of tiles: "rectangle", "tri1", "tri2".}

\item{group_prefix}{String prefix for polygon group IDs.}
}
\value{
Data frame with polygon vertices including \code{group} and \code{order} columns.
}
\description{
Takes a grid of points and tessellates into polygon tiles.
}
\keyword{internal}
