% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{position_param}
\alias{position_param}
\title{Position param}
\arguments{
\item{position}{Position adjustment, defaults to "identity". To collapse the result
onto one 2D surface, use \code{position_on_face()}.}
}
\description{
Position param
}
\keyword{internal}
