% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_data.R
\name{proj_data}
\alias{proj_data}
\title{Add projected data}
\usage{
proj_data(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Add projected data
}
