% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_gridlines.R
\name{proj_gridlines}
\alias{proj_gridlines}
\title{Add gridlines}
\usage{
proj_gridlines(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  faces = "open",
  color = "gray40",
  linewidth = 0.25,
  ...
)
}
\description{
Add gridlines
}
