% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_labels.R
\name{proj_label}
\alias{proj_label}
\title{Add axis labels}
\usage{
proj_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Add axis labels
}
