% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_margin.R
\name{proj_margin}
\alias{proj_margin}
\title{Add marginal data}
\usage{
proj_margin(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  faces = "far",
  ...
)
}
\description{
Add marginal data
}
