% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_mesh.R
\name{proj_mesh}
\alias{proj_mesh}
\title{Ridgeline layer}
\usage{
proj_mesh(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prj,
  ...
)
}
\description{
Ridgeline layer
}
