% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_panel.R
\name{proj_panel}
\alias{proj_panel}
\title{Add 3D cube faces}
\usage{
proj_panel(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  faces = "open",
  color = "black",
  fill = "gray80",
  linewidth = 0.25,
  ...
)
}
\description{
Add 3D cube faces
}
