% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_ridgeline.R
\name{proj_ridgeline}
\alias{proj_ridgeline}
\title{Ridgeline layer}
\usage{
proj_ridgeline(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Ridgeline layer
}
