% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_surface.R
\name{proj_surface}
\alias{proj_surface}
\title{3D surface}
\usage{
proj_surface(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prj,
  ...
)
}
\arguments{
\item{prj}{A \code{proj} object created by \code{projection()}.}
}
\value{
a ggplot layer
}
\description{
This stat visualizes a 3D surface by constructing a set of rectangular polygons and projecting them into 2D.
This function requires regular data, in which x and y coordinates form a grid, each combination of x and y
appears once, and the associated z value gives the height of the surface at that location.
}
\details{
The following \code{after_stat()} variables are available:
\itemize{
\item \code{after_stat(dzdx)}: the slope of the tile in the x-direction
\item \code{after_stat(dzdy)}: the slope of the tile in the y-direction
\item \code{after_stat(slope)}: the overall slope of the tile
\item \code{after_stat(aspect)}: the directional orientation of the tile
\item \code{after_stat(x)}: the x-coordinate after projection
\item \code{after_stat(y)}: the y-coordinate after projection
\item \code{after_stat(z)}: the z-coordinate after projection
}
}
\examples{
library(tidyverse)

# set up data and define projection
d <- expand_grid(x = seq(-10, 10, .5),
                 y = seq(-10, 10, .5)) \%>\%
      mutate(z = cos(sqrt(x^2 + y^2)))
prj <- projection(data = d, pitch = 35, yaw = 55,
      persp = T, dist = 1, zlim = c(-3, 3))

# surface plot, colored by z
ggplot(d, aes(x, y, z = z, fill = z)) +
      proj_surface(prj = prj) +
      scale_fill_viridis_c(option = "B") +
      theme_cube() +
      coord_fixed()

# highlight terrain attributes using after_stat
ggplot(d, aes(x, y, z = z, fill = after_stat(dzdy))) +
      proj_surface(prj = prj) +
      scale_fill_viridis_c(option = "B") +
      theme_cube() +
      coord_fixed()
}
