% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{Define a 3D plotting projection}
\usage{
projection(
  shear = 0,
  yaw = 0,
  pitch = 0,
  roll = 0,
  persp = F,
  dist = 1,
  hjust = 0.5,
  vjust = 0.5,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  zbreaks = NULL,
  data = NULL
)
}
\description{
This function creates a `proj` object containing the parameters needed to
project 3D data onto a 2D surface.
}
