% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-mixed.R
\name{render_single_type}
\alias{render_single_type}
\title{Dispatch to the correct renderer for a single primitive type}
\usage{
render_single_type(data, prim_type, rule, arrow, lineend)
}
\description{
Dispatch to the correct renderer for a single primitive type
}
\keyword{internal}
