% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_3d.R
\name{replace_colorbar_colors}
\alias{replace_colorbar_colors}
\title{Replace colorbar colors with shading gradients}
\usage{
replace_colorbar_colors(grob, grad)
}
\arguments{
\item{grob}{Original colorbar grob}

\item{grad}{Matrix of gradient colors}
}
\value{
Modified grob with gradients
}
\description{
Replace colorbar colors with shading gradients
}
\keyword{internal}
