% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{rgb2hsl}
\alias{rgb2hsl}
\title{Convert RGB to HSL color space}
\usage{
rgb2hsl(rgb_matrix)
}
\arguments{
\item{rgb_matrix}{3xN matrix with RGB values in \link{0,1} range}
}
\value{
3xN matrix with HSL values (H in \link{0,1}, S in \link{0,1}, L in \link{0,1})
}
\description{
Convert RGB to HSL color space
}
\keyword{internal}
