% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{rotate_3d}
\alias{rotate_3d}
\title{Rotate points in 3D space}
\usage{
rotate_3d(xyz, pitch, roll, yaw)
}
\arguments{
\item{xyz}{Matrix of x, y, z coordinates}

\item{pitch}{Rotation around x-axis in degrees}

\item{roll}{Rotation around y-axis in degrees}

\item{yaw}{Rotation around z-axis in degrees}
}
\value{
Rotated coordinates matrix
}
\description{
Rotate points in 3D space
}
\keyword{internal}
