% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-3d.R
\name{rotation_matrix_around_axis}
\alias{rotation_matrix_around_axis}
\title{Rotation matrix for rotation around an axis}
\usage{
rotation_matrix_around_axis(axis, angle)
}
\arguments{
\item{axis}{Unit vector defining rotation axis}

\item{angle}{Angle in degrees}
}
\value{
3x3 rotation matrix
}
\description{
Rotation matrix for rotation around an axis
}
\keyword{internal}
