% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-3d.R
\name{rotation_matrix_between_vectors}
\alias{rotation_matrix_between_vectors}
\title{Compute rotation matrix from one vector to another using Rodrigues' formula}
\usage{
rotation_matrix_between_vectors(from, to)
}
\arguments{
\item{from}{Unit vector to rotate from}

\item{to}{Unit vector to rotate to}
}
\value{
3x3 rotation matrix
}
\description{
Compute rotation matrix from one vector to another using Rodrigues' formula
}
\keyword{internal}
