% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{scale_to_npc_coordinates}
\alias{scale_to_npc_coordinates}
\title{Scale transformed coordinates to final npc coordinates}
\usage{
scale_to_npc_coordinates(x, y = NULL, plot_bounds)
}
\arguments{
\item{plot_bounds}{Plot bounds vector \link{xmin, xmax, ymin, ymax}}

\item{result}{Data frame with transformed x, y coordinates}
}
\value{
Data frame with scaled coordinates
}
\description{
Scale transformed coordinates to final npc coordinates
}
\keyword{internal}
