% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{scale_to_standard}
\alias{scale_to_standard}
\title{Scale data to standard domain \link{-0.5, 0.5}}
\usage{
scale_to_standard(values, data_range)
}
\arguments{
\item{values}{Vector of values to scale}

\item{data_range}{Original range of the data \link{min, max}}
}
\value{
Scaled values in \link{-0.5, 0.5} domain
}
\description{
Scale data to standard domain \link{-0.5, 0.5}
}
