% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{sort_params}
\alias{sort_params}
\title{Sorting parameters}
\arguments{
\item{sort_method}{Depth sorting algorithm. See \link{sorting_methods} for details.}
}
\description{
Sorting parameters
}
\keyword{internal}
