% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sphere_points}
\alias{sphere_points}
\title{Sphere surface points}
\format{
\subsection{\code{sphere_points}}{

A data frame with 200 rows and 3 columns (x, y, and z)
}
}
\usage{
sphere_points
}
\description{
A toy data set representing points on a sphere surface,
useful for testing hull functions
}
\keyword{datasets}
