% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{split_quads_to_triangles}
\alias{split_quads_to_triangles}
\title{Split quadrilateral tiles into triangles}
\usage{
split_quads_to_triangles(tile_vertices, grid_type)
}
\description{
Split quadrilateral tiles into triangles
}
\keyword{internal}
