% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube.R
\name{theme_cube}
\alias{theme_cube}
\title{A ggplot theme for ggcube}
\usage{
theme_cube(...)
}
\arguments{
\item{...}{arguments passed to \code{ggplot2::theme()}}
}
\value{
a gg theme object
}
\description{
This is a variant on \code{ggplot::theme_gray()} with blank panel background,
grid, and axis elements.
}
