% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-3d.R
\name{translate_face_names_from_flips}
\alias{translate_face_names_from_flips}
\title{Translate face names using stored flip information}
\usage{
translate_face_names_from_flips(face_names, scale_flips)

translate_face_names_from_flips(face_names, scale_flips)
}
\arguments{
\item{face_names}{Character vector of face names}

\item{faces}{Character vector of face names}

\item{flips}{List with x, y, z flip indicators}

\item{scale_x, scale_y, scale_z}{Scale objects for each axis}
}
\value{
Character vector of translated face names

Character vector of translated face names
}
\description{
Translate face names using stored flip information

Translate face names to account for scale direction flips
}
\keyword{internal}
