% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{compute_light_in_coord}
\alias{compute_light_in_coord}
\title{Compute lighting from standardized coordinates in coord_3d}
\usage{
compute_light_in_coord(data, standardized_coords, scale_ranges, scales, ratio)
}
\arguments{
\item{data}{Original data frame with lighting_spec column}

\item{standardized_coords}{Data frame with standardized x, y, z coordinates}
}
\value{
data frame with lighting values and normal components added
}
\description{
Compute lighting from standardized coordinates in coord_3d
}
\keyword{internal}
