% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{compute_triangle_normals}
\alias{compute_triangle_normals}
\title{Compute triangle normals from vertex coordinates}
\usage{
compute_triangle_normals(data, face_data)
}
\arguments{
\item{data}{Face vertex data}

\item{face_data}{Unique face data}
}
\value{
Matrix with normalized normal vectors (one row per face, 3 columns)
}
\description{
Compute triangle normals from vertex coordinates
}
\keyword{internal}
