% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-smooth-3d.R
\name{geom_smooth_3d}
\alias{geom_smooth_3d}
\title{3D smoothing geometry}
\usage{
geom_smooth_3d(
  mapping = NULL,
  data = NULL,
  stat = StatSmooth3D,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
3D smoothing geometry
}
