% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcube-package.R
\docType{package}
\name{ggcube}
\alias{ggcube-package}
\alias{ggcube}
\title{ggcube: 3D plotting extension for ggplot2}
\description{
ggcube extends ggplot2 into the third dimension, providing 3D coordinate
systems, surface plotting, and volumetric visualization capabilities.
}
\details{
To use ggcube, load both packages:
\code{library(ggplot2)}
\code{library(ggcube)}

Key functions:
\itemize{
\item \code{coord_3d()}: 3D coordinate system with rotation and perspective
\item \code{aes()}: Enhanced aesthetic mapping with positional z support
\item \code{stat_surface_3d()}: Surface plotting from grid data
\item \code{stat_hull_3d()}: 3D convex and alpha hulls
\item \code{stat_voxel_3d()}: 3D voxel visualization
\item \code{stat_pillar_3d()}: 3D pillar/bar charts
\item \code{light()}: Lighting specifications for 3D surfaces
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewkling/ggcube}
  \item \url{https://matthewkling.github.io/ggcube/}
  \item Report bugs at \url{https://github.com/matthewkling/ggcube/issues}
}

}
\keyword{internal}
