% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{light_param}
\alias{light_param}
\title{Light param}
\arguments{
\item{light}{A lighting specification object created by \code{\link[=light]{light()}}, or NULL to disable shading.}
}
\description{
Light param
}
\keyword{internal}
