% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{light_param}
\alias{light_param}
\title{Light param}
\arguments{
\item{light}{A lighting specification object created by \code{\link[=light]{light()}},\code{"none"} to disable
lighting, or \code{NULL} to inherit plot-level lighting specs from the coord. Specify
plot-level lighting in \code{coord_3d()} and layer-specific lighting in \verb{geom_*3d()}
functions.}
}
\description{
Light param
}
\keyword{internal}
