% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-gen.R
\name{make_point_grid}
\alias{make_point_grid}
\title{Generate a grid of unique vertex points}
\usage{
make_point_grid(n = 40, xlim, ylim)
}
\arguments{
\item{n}{Integer or length-2 integer vector specifying grid resolution.}

\item{xlim, ylim}{Length-two numeric vectors defining bounding box.}
}
\value{
A data frame with columns \code{x}, \code{y}, \code{row}, \code{column}.
}
\description{
Creates a regular grid of unique vertex positions. Unlike \code{\link[=make_tile_grid]{make_tile_grid()}},
this returns one row per unique vertex position rather than polygon vertices
with duplicates.
}
\keyword{internal}
