% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{polygon_params}
\alias{polygon_params}
\title{Parameters for layers using GeomPolygon3D}
\arguments{
\item{cull_backfaces, sort_method, force_convex, scale_depth}{Advanced polygon
rendering parameters. See \link{polygon_rendering} for details.}

\item{na.rm}{If \code{FALSE}, missing values are removed.}

\item{show.legend}{Logical indicating whether this layer should be included in legends.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics.}

\item{...}{Other arguments passed on to the the layer function (typically GeomPolygon3D), such as
aesthetics like \code{colour}, \code{fill}, \code{linewidth}, etc.}
}
\description{
Parameters for layers using GeomPolygon3D
}
\keyword{internal}
