% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{polygon_rendering}
\alias{polygon_rendering}
\title{Polygon rendering parameters}
\arguments{
\item{cull_backfaces}{Logical indicating whether to remove back-facing polygons
from rendering for performance optimization. Back faces are determined using
screen-space winding order after 3D transformation. Defaults vary by geometry
type: TRUE for solid objects (hulls, voxels), FALSE for surfaces and wireframes.}

\item{sort_method}{Character indicating algorithm used to determine the order in which
polygons are rendered.
\itemize{
\item \code{"painter"}: Polygons are sorted by the mean depth (distance from viewer after
rotation) of their vertices. This is fast, but can give incorrect results in certain cases.
\item \code{"pairwise"}: A more intensive sorting algorithm that compares every pair of
polygons in 3D to determine which face should be rendered behind the other;
slower but more accurate.
\item \code{"auto"}: The default. Uses pairwise if polygon data has less
than 500 rows and painter otherwise.
}}

\item{scale_depth}{Logical indicating whether polygon linewidths should be scaled to make closer lines
wider and farther lines narrower. Default is TRUE. Scaling is based on the mean depth of a polygon.}

\item{force_convex}{Logical indicating whether to remove polygon vertices that are not part of the
convex hull. Default is FALSE. Specifying TRUE can be useful for preventing artifacts in surfaces
that have polygon tiles that wrap over a visible horizon.}
}
\description{
Polygon rendering parameters
}
