% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_surface.R
\name{proj_surface}
\alias{proj_surface}
\title{Surface}
\usage{
proj_surface(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prj,
  ...
)
}
\arguments{
\item{prj}{A `proj` object created by \code{projection()}.}
}
\value{
a ggplot
}
\description{
Surface
}
\examples{
library(tidyverse)

# set up data and define projection
d <- expand_grid(x = seq(-10, 10, .5),
                 y = seq(-10, 10, .5)) \%>\%
      mutate(z = cos(sqrt(x^2 + y^2)))
prj <- projection(data = d, pitch = 35, yaw = 55,
      persp = T, dist = 1, zlim = c(-3, 3))

# surface plot, colored by z
ggplot(d, aes(x, y, z = z, fill = z)) +
      proj_surface(prj = prj) +
      scale_fill_viridis_c(option = "B") +
      theme_cube() +
      coord_fixed()

# highlight terrain attributes using after_stat
ggplot(d, aes(x, y, z = z, fill = after_stat(dzdy))) +
      proj_surface(prj = prj) +
      scale_fill_viridis_c(option = "B") +
      theme_cube() +
      coord_fixed()
}
